Added lots of tests

+++ src/main/kotlin/ch/frostnova/tools/kafkarouter/util/BackoffStrategy.kt
@@ -4,7 +4,7 @@ import java.lang.Thread.sleep
 
 class BackoffStrategy(
     val retryCount: Int = Int.MAX_VALUE,
-    val backoffTimeSeconds: List<Int> = listOf(1, 2, 3, 5, 10, 30)
+    val backoffTimeSeconds: List<Double> = listOf(1.0, 2.0, 3.0, 5.0, 10.0, 30.0)
 ) {
 
     val logger = logger(BackoffStrategy::class)
@@ -23,12 +23,12 @@ class BackoffStrategy(
                 lastException = ex
                 val nextRetryOffset =
                     if (count < backoffTimeSeconds.size) backoffTimeSeconds[count] else backoffTimeSeconds.lastOrNull()
-                        ?: 0
+                        ?: 0.0
                 count++
                 if (count < retryCount) {
                     logger.warn("caught ${ex.javaClass.simpleName}: ${ex.message} from ${ex.stackTrace[1]}, retry #$count in $nextRetryOffset seconds...")
                 }
-                sleep(nextRetryOffset * 1000L)
+                sleep((nextRetryOffset * 1000L).toLong())
             }
         }
         throw lastException ?: IllegalStateException()
